274: Fixed review issues
