+++ src/helpers/canvas-helpers.ts
@@ -18,3 +18,18 @@ export function strokeRectInnerWithFill(ctx: CanvasRenderingContext2D, x: number
 	// right
 	ctx.fillRect(x + w - lineWidth, y, lineWidth, h);
 }
+
+export function drawScaled(ctx: CanvasRenderingContext2D, ratio: number, func: () => void): void {
+	ctx.save();
+	ctx.scale(ratio, ratio);
+	func();
+	ctx.restore();
+}
+
+export function clearRect(ctx: CanvasRenderingContext2D, x: number, y: number, w: number, h: number, clearColor: string): void {
+	ctx.save();
+	ctx.globalCompositeOperation = 'copy';
+	ctx.fillStyle = clearColor;
+	ctx.fillRect(x, y, w, h);
+	ctx.restore();
+}
